move helper functions to helpers folder
